Look-up of methods working.
Doesn't matter if the target method is static, private, abstract or from
an interface, the compiler won't complain, and yet, check for existence
of the class and will try to best-match the methods based on the
signature.

This required not using the invocation-checking-code, but changing a few
methods to receive a 'checkAccess'. Not sure if we are loading
everything we need, though.