Subtype annotation for polymorphic callsites.
Now we can do the following:

  propagate E: A::a() -> {B,C <: D}::d() -> Z::z();

Where any call D.d() will branch on B.d() and C.d() when
building the callgraph (altough it still doesn't check if B
and C are subtypes of D -- but d() must be own-method of
those classes; not inherited).

Overriding is still followed on simple nodes (but not for
polyp-nodes). So, if z() is not a Z method, we will take the
inherited z and proceed.