Changing the meaning of polymorphic nodes.
For convenience while working on freemind's codebase, I
decided to change the meaning of {A,B<:C} to explore all
paths specified (ie. when a method call is found, instead of
matching it's receiver type to the base C, match it to any
of the types specified). The "<:" operator remains important
to control to the overriding logic.

   # Please enter the commit message for your changes. Lines starting