improvements on polymorphic nodes (with original semantics)
Given {A,B,C<:D}:
o checks if A, B and C are subclass of D.

o Sort A,B and C them by distance to D, so the compiler
 doesn't raise an error because of ordering issues.

Also, the original semantics of polymorphic nodes was
restored: when constructing the callgraph, the node is only
matched on the supertype (ie. the callsite's type is not
matched on subtypes anymore).