Huge refactoring of the scripts to deal with polymorphism.
Overriding nodes generate hierarchy information (type/subtype
relationships). Then, for every node in the callgraph whose type is a
supertype of a previous known node, we will use the full polymorphic
node in each site. When no such node is part of a path, but we need to
tell the compiler that the overriding goes up in the hierarchy, we mark
a one-node path with the polymorphic node and prepend it to the list of
propagating statements.