Improved checking of methods and parameters (+generics).
Refactored dirtyAttrib() in Attr and placed a version of it
in MemberEnter. Now the appropriated environment to evaluate
each type (so as to workaround access policy) is obtained
with "enter.typeEnvs.get(Symbol)".

Access to private methods and private types should be
working and parameterized types are partially working: type
T is correctly evaluated in its context, but instantiation
of types on inheritance brings an issue:

class A<T> {
  void a(T t) { ...}
}

class B extends A<String> {
  void a(String t) { ...}
}

propagating E: {B<:A}::a(T); //raises error